Site Settings: Move DeleteSite, StartOver, ThemeSetup and DateTimeFormat to General section
