+++ client/my-sites/site-settings/index.js
@@ -24,11 +24,5 @@ module.exports = function() {
 		page( '/settings/export/:site_id', controller.siteSelection, controller.navigation, settingsController.exportSite );
 	}
 
-	if ( config.isEnabled( 'manage/site-settings/delete-site' ) ) {
-		page( '/settings/delete-site/:site_id', controller.siteSelection, controller.navigation, settingsController.setScroll, settingsController.deleteSite );
-		page( '/settings/start-over/:site_id', controller.siteSelection, controller.navigation, settingsController.setScroll, settingsController.startOver );
-		page( '/settings/theme-setup/:site_id', controller.siteSelection, controller.navigation, settingsController.setScroll, settingsController.themeSetup );
-	}
-
 	page( '/settings/:section', settingsController.legacyRedirects, controller.siteSelection, controller.sites );
 };
